==> Source code/02_sh10050_MCF7_ssGSEA.Rmd <==
---
title: "sh10050 MCF7 ssGSEA"
output: html_document
date: "2024-01-30"
editor_options: 
  chunk_output_type: console
---

```{r}
library(data.table)
library(openxlsx)
library(foreach)
library(conflicted)
library(msigdbr)
library(GSEABase)
library(GSVA)
library(tidyverse)
conflict_prefer_all("dplyr", quiet = T)

dir <- "~/u01_rna.seq/sh10050/"
c.line <- "MCF7"
batch <- "sh10050"

#Perform ssGSEA
```


```{r}
set.seed(1111)

#read in log2 transfromed counts
setwd(dir = paste0(dir, "counts/filtered"))
lcpm.filt <- fread(file = "sh10050_logCPM_filtered_normalized_13838_genes.txt",
                   sep = "\t", header = T, quote = "")
lcpm.filt <- lcpm.filt %>% 
  as.data.frame() %>% 
  colnames_to_rownames(var= "gene") %>% 
  as.matrix()

collection <- c("H", "C2", "C5", "C6")

gsva. <- list()

#Run ssGSEA
for (i in collection) {
  
  #Get pathways for collection
  pathways <- msigdbr(species = "Homo sapiens", category = i) %>%
  group_by(gs_name)
  pathways <- split(x = pathways$gene_symbol, f = pathways$gs_name)
  
  #Run ssgsea
  gsva.[[i]] <- gsva(expr = lcpm.filt, gset.idx.list =  pathways,
                method="ssgsea",  min.sz = 15, max.sz = 500, verbose=TRUE, 
                kcdf="Gaussian", parallel.sz=1L, mx.diff = T,
                tau = 0.25, ssgsea.norm = T)
}

#Split list based on collection
gsva.H <- gsva.$H %>% 
  as.data.frame() %>%
  rownames_to_column(var = "Gene_Set")
gsva.c2 <- gsva.$C2 %>% 
  as.data.frame() %>%
  rownames_to_column(var = "Gene_Set")
gsva.c5 <- gsva.$C5 %>% 
  as.data.frame() %>%
  rownames_to_column(var = "Gene_Set")
gsva.c6 <- gsva.$C6 %>% 
  as.data.frame() %>%
  rownames_to_column(var = "Gene_Set")

gsva.2 <- list("H" = gsva.H, 
               "C2" = gsva.c2, 
               "C5" = gsva.c5, 
               "C6" = gsva.c6)


setwd(dir = paste0(dir, "ssgsea/raw_output/"))
openxlsx::write.xlsx(gsva.2, 
                     file = paste(batch, c.line,
                                   "ssgsea_H.C2.C5.C6",
                                   "ssGSEA_scores.xlsx", sep = "_"))
setwd(paste0(dir, "ssgsea/raw_output/"))
saveRDS(gsva.2, file = paste(batch, c.line, "ssgsea_H.C2.C5.C6",
                             "ssGSEA_scores.rds", sep = "_"))


#Make dataframe for statistical testing for all collections

#i <- "H"
collection <- c("H", "C2", "C5", "C6")
#collection <- "C6"
#gsva.3.list <- list()

start <- Sys.time()
gsva.3.list <- foreach(i = collection) %dopar% {
  print(i)
  
  gsva.2[[i]] %>% 
    as.data.frame() %>%
    mutate("1_MCF7_Sen_DMSO_D0H0" = .[["1_MCF7_Sen_NT_D0H0"]],
         "1_MCF7_Sen_Afat_D0H0" = .[["1_MCF7_Sen_NT_D0H0"]],
         "1_MCF7_Sen_Ribo_D0H0" = .[["1_MCF7_Sen_NT_D0H0"]],
         "1_MCF7_Sen_Comb_D0H0" = .[["1_MCF7_Sen_NT_D0H0"]],
         "2_MCF7_RiboR_DMSO_D0H0" = .[["2_MCF7_RiboR_NT_D0H0"]],
         "2_MCF7_RiboR_Afat_D0H0" = .[["2_MCF7_RiboR_NT_D0H0"]],
         "2_MCF7_RiboR_Ribo_D0H0" = .[["2_MCF7_RiboR_NT_D0H0"]],
         "2_MCF7_RiboR_Comb_D0H0" = .[["2_MCF7_RiboR_NT_D0H0"]]) %>% 
  dplyr::select(-c("1_MCF7_Sen_NT_D0H0", "2_MCF7_RiboR_NT_D0H0")) %>% 
  pivot_longer(! Gene_Set, names_to = "Cell", 
               values_to = "Enrichment_Score") %>%
  as.matrix() %>%
  as.data.frame() %>%
  separate(Cell, into = c("num", "CellLine", "State", "Treatment", "Time"),
           sep = "_", remove = F) %>%
  unite(S_T, c(State, Treatment), sep = "_", remove = F) %>%
  unite(S_T_T, c(State, Treatment, Time), sep = "_", remove = F) %>%
  dplyr::select(Gene_Set, Enrichment_Score, Cell, CellLine, S_T, S_T_T, 
                State, Treatment, Time, num) %>% 
  mutate(Hour = ifelse(Time == "D0H0", 0,
                ifelse(Time == "D0H6", 6,
                ifelse(Time == "D0H24", 24,
                ifelse(Time == "D6H0", 144,
                ifelse(Time == "D6H6", 150,
                ifelse(Time == "D6H24", 168,
                ifelse(Time == "D12H0", 288,
                ifelse(Time == "D12H6", 294,
                ifelse(Time == "D12H24", 312,
                ifelse(Time == "D18H0", 432,
                ifelse(Time == "D18H6", 438,
                ifelse(Time == "D18H24", 456, ""))))))))))))) %>% 
  mutate_at(vars(Gene_Set, CellLine, Cell, S_T, S_T_T, Treatment, State, Time),
            as.factor) %>% 
  mutate_at(vars(Enrichment_Score), as.numeric) %>% 
    
  mutate_at(vars(Hour), as.integer) %>% 
  mutate(Treatment = relevel(Treatment, ref = "DMSO"))  

}
names(gsva.3.list) <- collection
end <- Sys.time()
df.stats <- paste("loop took ", end - start)
print(df.stats)

#Save Scores
setwd(dir = paste0(dir, "ssgsea/raw_output/"))
saveRDS(gsva.3.list, file = paste("ssGSEA_collections_H_C2_C5_C6_scores_list.rds")
```
